UAPI: ndctl: Remove dependency on PAGE_SIZE macro
The PAGE_SIZE macro should not be used in UAPI headers because it is not
guaranteed to be present outside of the kernel. Hard-code the page size
instead to fix compilation in edge cases where the macro is unavailable
in userspace.

This change was extracted from Microsoft's official May 2019 WSL2 kernel
snapshot: https://thirdpartysource.microsoft.com/download/Windows%20Subsystem%20for%20Linux%20v2/May%202019/WSLv2-Linux-Kernel-master.zip

Signed-off-by: Danny Lin <danny@kdrag0n.dev>